deprecate NoPersistHttpDataAccess + handle no persistence inside DataAccess class
