+++ packages/request-client.js/src/http-request-network.ts
@@ -12,7 +12,6 @@ import RequestNetwork from './api/request-network';
 import HttpDataAccess, { NodeConnectionConfig } from './http-data-access';
 import { MockDataAccess } from '@requestnetwork/data-access';
 import { MockStorage } from './mock-storage';
-import { NoPersistHttpDataAccess } from './no-persist-http-data-access';
 
 /**
  * Exposes RequestNetwork module configured to use http-data-access.
@@ -60,12 +59,7 @@ export default class HttpRequestNetwork extends RequestNetwork {
       ? new MockDataAccess(new MockStorage(), {
           persist: !skipPersistence,
         })
-      : skipPersistence
-      ? new NoPersistHttpDataAccess({
-          httpConfig,
-          nodeConnectionConfig,
-        })
-      : new HttpDataAccess({ httpConfig, nodeConnectionConfig, persist: true });
+      : new HttpDataAccess({ httpConfig, nodeConnectionConfig, persist: !skipPersistence });
 
     if (!currencyManager) {
       currencyManager = CurrencyManager.getDefault();
